ConstructionSites - Development Commit 0.9.0
• Added build stages.

Make the building, then make the stages for it. To make stages, just
save a schematic that looks like an unfinished version with the time
that you'd like it to be shown at in the title in a special manner.
Example: first create a build site named, say, "treehouse" the way you
normally would, then create a schematic named "treehouse_0-2-30" and
save it. At two minutes and thirty seconds, that schematic will be
placed.

• Changes to how scanning is handled.

• Other small improvements.